refactor(desktop): centralize navigation ports
Generated-by: OpenAI Codex
