+++ apps/desktop/src/renderer/features/session-navigation/index.ts
@@ -21,11 +21,10 @@ export { SessionNavigationServicesProvider } from './services-context.js';
 export { SessionNavigationProvider } from './ui/session-navigation-provider.js';
 export { createSessionOpenCommand } from './controller/session-open-command.js';
 export { useSessionNavigationReads } from './controller/use-session-navigation-reads.js';
-export type { SessionNavigationPorts } from './controller/use-session-navigation-controller.js';
 export { deriveSessionRail } from './model/session-rail.js';
 export { sessionRailLayoutStore } from './model/session-rail-layout-store.js';
 export type {
   SessionNavigationRowActions,
   SessionPurgeOutcome,
 } from './controller/session-row-actions.js';
-export type { SessionNavigationServices } from './ports.js';
+export type { SessionNavigationPorts, SessionNavigationServices } from './ports.js';
